Consolidate garbage collection
Currently, old snapshots and xlogs are deleted by the snapshot daemon
while vinyl files are removed by the Engine::waitCheckpoint callback.
For the sake of backups and replication, which need to temporarily
disable garbage collection, we should bring all garbage collection
routines together in one place. So this patch introduces box.gc()
method, which takes an LSN of the latest snapshot to keep as its
argument. When called, it deletes all xlog files as well as engine
specific files (memtx snapshots, vinyl runs) that are not required to
recover from a snapshot with LSN greater or equal to the given one.
For removal of engine specific files, a new engine callback is
introduced, Engine::collectGarbage. The snapshot daemon now calls
box.gc() to cleanup instead of deleting snap and xlog files by itself.